Fix more warnings

+++ src/plugins/elev.c
@@ -114,7 +114,7 @@ static guint16 *_load_bil(gchar *path)
 	g_debug("GritsPluginElev: load_bil %p", data);
 	if (len != TILE_SIZE) {
 		g_warning("GritsPluginElev: _load_bil - unexpected tile size %ld, != %ld",
-				len, TILE_SIZE);
+				(glong)len, (glong)TILE_SIZE);
 		g_free(data);
 		return NULL;
 	}
